+++ collective/portlet/carousel/browser.py
@@ -11,6 +11,50 @@ from plone.memoize.view import memoize
 
 from z3c.form import field
 
+script = """
+jQuery(function($){
+  rotate = %(rotate)s;
+  var timeout;
+
+  next = function(index, auto) {
+    clearTimeout(timeout);
+    x = parseInt(index);
+    p = $('.portletCarousel ul li');
+    i = $('.portletCarousel dl dd.carouselPortletItem');
+    p.removeClass('current');
+    i.removeClass('current').animate({opacity:0});
+    i.hide();
+    x%%=i.length;
+    ci = i.eq(x);
+    cp = p.eq(x);
+    cp.addClass('current');
+    ci.addClass('current').animate({opacity:1});
+    ci.show();
+    if(auto) {
+      timeout = setTimeout("next("+(++x)+", " + auto + ")", %(timeout)d);
+    }
+  }
+
+  $('.portletCarousel ul li').click(function(){
+    cp = $(this);
+    if (cp.is('.current'))
+      return false;
+    x = cp.index('.portletCarousel ul li');
+    next(x, true);
+    return false;
+  });
+
+  if (rotate) {
+    $(document).ready(function() {
+      timeout = setTimeout("next(1,true)", %(timeout)d);
+    });
+  }
+
+});
+
+
+"""
+
 
 class CarouselPortletRenderer(base.Renderer):
     render = ViewPageTemplateFile('portlet.pt')
@@ -31,7 +75,11 @@ class CarouselPortletRenderer(base.Renderer):
     @property
     def rotate(self):
         return self.data.automatic_rotation and \
-            len(self.items()) > 1
+            len(self.items()) > 1 and self.timeout > 0
+
+    @property
+    def timeout(self):
+        return self.data.timeout * 1000
 
     @memoize
     def items(self):
@@ -55,6 +103,10 @@ class CarouselPortletRenderer(base.Renderer):
             return items[:self.data.limit]
 
         return items
+
+    def script(self):
+        return script % dict(rotate=self.rotate and 'true' or 'false',
+                             timeout=self.timeout)
 
 
 class CarouselPortletAddForm(z3cformhelper.AddForm):